requirements.txt
Flask-SQLAlchemy
psycopg2
pytest-cov